requirements.txt
flask
sqlalchemy
flask_bcrypt
flask_sqlalchemy
flask_wtf
flask_login
flask_paginate
pymysql
psycopg2-binary
pytz
pyyaml
mysqlclient